requirements.production.txt
Django>=5.2,<6.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
PyJWT>=2.8.0
cryptography>=41.0.0
django-allauth>=0.57.0
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
sendgrid>=6.11.0

# Security packages
django-debug-toolbar>=4.2.0
django-cors-headers>=4.3.0

# For HTTPS
certifi>=2023.11.17 